move tests to dedicated test class
