+++ src/test/java/org/openrewrite/apache/httpclient5/UpgradeApacheHttpClient5Test.java
@@ -599,62 +599,4 @@ class UpgradeApacheHttpClient5Test implements RewriteTest {
           )
         );
     }
-
-    @Test
-    void setContentEncodingToConstructorArg() {
-        rewriteRun(
-          //language=java
-          java(
-            """
-              import org.apache.http.entity.ContentType;
-              import org.apache.http.entity.StringEntity;
-
-              class A {
-
-                  private void a() {
-                      StringEntity se = new StringEntity("", ContentType.APPLICATION_JSON);
-                      se.setContentEncoding("utf-8");
-                  }
-              }
-              """,
-            """
-              import org.apache.hc.core5.http.ContentType;
-              import org.apache.hc.core5.http.io.entity.StringEntity;
-
-              class A {
-
-                  private void a() {
-                      StringEntity se = new StringEntity("", ContentType.APPLICATION_JSON, "utf-8", false);
-                  }
-              }
-              """
-          )
-        //   java(
-        //     """
-        //       import org.apache.http.entity.ContentType;
-        //       import org.apache.http.entity.StringEntity;
-        //       import org.apache.http.message.BasicHeader;
-
-        //       class A {
-
-        //           private void a() {
-        //               StringEntity se = new StringEntity("", ContentType.APPLICATION_JSON);
-        //               se.setContentEncoding(new BasicHeader("Content-Encoding", "utf-8"));
-        //           }
-        //       }
-        //       """,
-        //     """
-        //       import org.apache.hc.core5.http.ContentType;
-        //       import org.apache.hc.core5.http.io.entity.StringEntity;
-
-        //       class A {
-
-        //           private void a() {
-        //               StringEntity se = new StringEntity("", ContentType.APPLICATION_JSON, "utf-8", false);
-        //           }
-        //       }
-        //       """
-        //   )
-        );
-    }
 }
